dockercompose: move config yaml parsing into the dockercompose package
